Fix servo assignments
In our original servo assignments, two of the CPN servos were set to act
as servo 7. This clashed with the PWM outputs to the PLB (for which
servo 7 and servo 14 are routed from the cube).  To fix this, we meant
to reassign Servo 7 to Servo 11.

I accidentally reassigned Servo 6 instead of Servo 7

SW-181