Common Defaults: enable terrain in Auto and Guided
Without this, Auto and Guided ignore terrain-frame commands. This does
NOT force terrain following on for relative-frame commands.

At the time of writing, Mission Planner only sends relative frame for
Guided mode, but this parameter change means that I can fix that in the
future without needing to update the firmware.

SW-159