hwdef: Volanti defaults: reduce filtering
We have multiple dynamic notches with deep attenuation. We have been
hitting some edge cases where the notch filters are locking onto low
frequencies and causing issues/crashes. We are reducing the filtering
to the bare minimum: reducing attenuation and bandwidth as much as
possible.

This also required a reduction in our rate gains.

SW-456