SITL: add Ottano.json model
In theory this file can be used to add physics information about your
aircraft, but the plane physics are stuck hard-coded as a medium foamy.
This is only used to get our battery simulation working for now, but in
the future, I will probably expand this to allow it to override the
forward-flight physics.