CarbonixCommon: BRD_SAFETY_MASK enable 10-15
This was missed when Ottano separated servo assignments to
one-per-physical-servo. Also, servo 14 (payload servo) should have been
included in the mask originally, which was done  in Volanti, but not
Ottano. For that reason, this mask was moved to common, even though the
servo assignments are different. This is safe because one platform will
never have a motor assigned where another has a servo.