CarbonixCommon: make QAssist more sensitive
The default of 30 degrees and half-second delay are too much. There's
already enough delay in the time it takes to start up the ESCs, and
the requirement of being 5 degrees out of envelope in addition to the
angle error threshold is enough to prevent false positives.

I would rather work on relaxing this going forward, erring on the side
of nuisance triggers than late triggers.